Move from domain.entity to entities
