+++ src/main/java/br/com/fiap/soat/techChallenge/application/adapter/inbound/web/controller/ProdutoController.java
@@ -2,7 +2,7 @@ package br.com.fiap.soat.techChallenge.application.adapter.inbound.web.controlle
 
 import br.com.fiap.soat.techChallenge.application.adapter.inbound.web.controller.model.response.ProdutoResponse;
 import br.com.fiap.soat.techChallenge.application.adapter.inbound.web.controller.model.request.ProdutoRequest;
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 import br.com.fiap.soat.techChallenge.domain.ports.inbound.CadastrarProdutoUseCasePort;
 import br.com.fiap.soat.techChallenge.domain.ports.inbound.EditarProdutoUseCasePort;
 import br.com.fiap.soat.techChallenge.domain.ports.inbound.ExcluirProdutoUseCasePort;
