+++ src/main/java/br/com/fiap/soat/techChallenge/application/adapter/outbound/repository/ProdutoRepositoryAdapter.java
@@ -2,7 +2,7 @@ package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository;
 import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.ProdutoEntity;
 import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.ProdutoJpaRepository;
 import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.mappers.ProdutoMapper;
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 import br.com.fiap.soat.techChallenge.domain.exceptions.ProdutoNaoEncontradoException;
 import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
 import jakarta.transaction.Transactional;
