+++ src/main/java/br/com/fiap/soat/techChallenge/application/adapter/outbound/repository/jpa/ProdutoJpaRepository.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa;
 import br.com.fiap.soat.techChallenge.application.adapter.outbound.repository.jpa.entity.ProdutoEntity;
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
