+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/CadastrarClienteUseCasePort.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
 
-import br.com.fiap.soat.techChallenge.domain.entity.Cliente;
+import br.com.fiap.soat.techChallenge.entities.Cliente;
 
 public interface CadastrarClienteUseCasePort {
     Cliente execute(Cliente cliente);
