+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/EditarProdutoUseCasePort.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
 
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 
 public interface EditarProdutoUseCasePort {
     Produto execute(Produto produto);
