+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/FazerPedidoUseCasePort.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
 
 import br.com.fiap.soat.techChallenge.domain.usecases.model.ComandoDeNovoPedido;
-import br.com.fiap.soat.techChallenge.domain.entity.Pedido;
+import br.com.fiap.soat.techChallenge.entities.Pedido;
 
 public interface FazerPedidoUseCasePort {
     Pedido execute(ComandoDeNovoPedido comandoDeNovoPedido);
