+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/IdentificarClienteUseCasePort.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
 
-import br.com.fiap.soat.techChallenge.domain.entity.Cliente;
+import br.com.fiap.soat.techChallenge.entities.Cliente;
 
 import java.util.Optional;
