+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/ObterPedidosUseCasePort.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
 
-import br.com.fiap.soat.techChallenge.domain.entity.Pedido;
+import br.com.fiap.soat.techChallenge.entities.Pedido;
 
 import java.util.List;
