+++ src/main/java/br/com/fiap/soat/techChallenge/domain/ports/inbound/ObterProdutosPorCategoriaUseCasePort.java
@@ -1,5 +1,5 @@
 package br.com.fiap.soat.techChallenge.domain.ports.inbound;
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 
 import java.util.List;
