+++ src/main/java/br/com/fiap/soat/techChallenge/domain/usecases/EditarProdutoUseCase.java
@@ -1,6 +1,6 @@
 package br.com.fiap.soat.techChallenge.domain.usecases;
 
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 import br.com.fiap.soat.techChallenge.domain.ports.inbound.EditarProdutoUseCasePort;
 import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
