+++ src/main/java/br/com/fiap/soat/techChallenge/domain/usecases/ObterProdutosPorCategoriaUseCase.java
@@ -1,5 +1,5 @@
 package br.com.fiap.soat.techChallenge.domain.usecases;
-import br.com.fiap.soat.techChallenge.domain.entity.Produto;
+import br.com.fiap.soat.techChallenge.entities.Produto;
 import br.com.fiap.soat.techChallenge.domain.ports.outbound.ProdutoRepositoryPort;
 import br.com.fiap.soat.techChallenge.domain.ports.inbound.ObterProdutosPorCategoriaUseCasePort;
